Added VERY BASIC menu functionality (Gamemaker)
Included rooms:    gamestart, mainMenu, loadGameMenu,Options,Menu

Menu has basic functionality:  main menu has a button for each room and
each room is equipped with a back button sending the user back to the
main menu.

Each room is also equipped with a room identifier to tell us which room
we are in.