Updated profiles and worked on load system
added the ability to delete profiles which also deletes all saves
associated with the profile.

started work on the load system... currently there isnt a load system
and it is just buttons that read what will be a saved game to see if
there is a save or not... this might take some time to get working.